move span to its own module
